casting lu

+++ src/Utils/printUtils.c
@@ -19,7 +19,7 @@ void print_symbol_table(Symbol **Symbols, size_t *symbol_count, const char *file
 
     dest = fopen(fullFileName, "w");
 
-    printf("\nPrinting symbol table with %zu symbols\n", *symbol_count);
+    printf("\nPrinting symbol table with %d symbols\n", (int)*symbol_count);
 
     for (i = 0; i < *symbol_count; i++)
     {
